refactor(daemon): move Slack turn output into its platform module (S2 §7.3)
Slack was the first platform, so its output code never had to justify itself
as platform-specific — it simply was "how the daemon posts". The other three
extractions found a self-contained applier and lifted it; here the applier had
grown seven private helpers in daemon.ts that read like core utilities but are
Slack API shapes end to end: chat.postMessage option bags, block-edit retries,
a chat.delete for the status bar.

The helpers move WITH the applier rather than becoming host capabilities. Had
they stayed, Slack's host port would be eight functions wide against Telegram's
two, and every one of the eight would take or return a Slack type — a port in
name only. Two groups move:

  - Applier-private: postSlackReply (the born-with-its-footer post boundary),
    updateSlackLiveReply, findExistingSlackStatusBarTs.
  - Shared with core call sites: the four option builders,
    clearStaleSlackReplyFooters, and isSlackStatusBarText. These are pure
    functions of the turn record, so they move as exported functions and core
    imports them back — the dependency now points core → platform.

SlackTurnState (staleReplyFooters) moves to the module that owns it, joining
the Telegram / Feishu state shapes already extracted.

What deliberately did NOT move: liveReplyTs / progressTs / statusBarTs /
lastReply / attribution and their *Attempted flags stay on the core turn
record. They look like Slack state, and most of them belong in the opaque slot
eventually — but core reads them today from suppression, re-anchoring, and
human-input paths (attribution alone has 30+ readers). The structural SlackTurn
view names that surface exactly, which turns the follow-up into a mechanical
field migration. Relocating a body while silently redefining what it may reach
is the §16 failure mode this refactor was written to avoid.

This applier is also the registry's CORE surface: webchat / hook / dream render
through it (§12). The `platform !== 'slack'` guards in the option builders are
what keep that safe — a non-Slack origin gets undefined options and posts
without Slack identity decoration.

daemon.ts: -404 / +41.
